Restructure around overridable
